svg/painting/parsing/marker-start-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Geometry Properties: getComputedValue().markerStart</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#MarkerStartProperty"/>
    <h:meta name="assert" content="marker-start computed value is as specified, with url values absolute."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("marker-start", "none");
test_computed_value("marker-start", 'url("https://example.com/")');

test(() => {
  const target = document.getElementById('target');
  target.style['marker-start'] = 'url("a.b#c")';
  const result = getComputedStyle(target)['marker-start'];
  assert_true(result.includes('/a.b#c'), result + ' is absolute');
}, 'url values are made absolute');

  ]]></script>
</svg>
