LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/marker-start-computed-expected.txt


PASS Property marker-start value 'none' 
FAIL Property marker-start value 'url("https://example.com/")' assert_equals: expected "url(\"https://example.com/\")" but got "none"
FAIL url values are made absolute assert_equals: expected "url(\"http://localhost:8800/svg/painting/parsing/a.b#c\")" but got "none"
